+++ org-roam-node.el
@@ -761,26 +761,34 @@ Assumes that the cursor was put where the link is."
 
 ;;;;;; Completion-at-point interface
 (defconst org-roam-bracket-completion-re
-  "\\[\\[\\(\\(?:roam:\\)?\\)\\([^z-a]*?\\)]]"
+  "\\[\\(?:\\[id:\\([^z-a]*?\\)]\\)?\\[\\([^z-a]+?\\)]]"
   "Regex for completion within link brackets.
-We use this as a substitute for `org-link-bracket-re', because
-`org-link-bracket-re' requires content within the brackets for a match.")
+Matches both empty links (i.e. \"[[|]]\") and existing \"id:\"
+links (e.g. \"[[id:12345][|]]\").")
 
 (defun org-roam-complete-link-at-point ()
-  "Complete \"roam:\" link at point to an existing Org-roam node."
+  "Complete \"id:\" link at point to an existing Org-roam node.
+Places the completed string into the link description. If there
+was already an \"id:\" link at point, overwrites it to fit the
+new description."
   (let (roam-p start end)
     (when (org-in-regexp org-roam-bracket-completion-re 1)
-      (setq roam-p (not (string-blank-p (match-string 1)))
+      (setq old-id (match-string 1)
             start (match-beginning 2)
             end (match-end 2))
       (list start end
             (org-roam--get-titles)
             :exit-function
             (lambda (str &rest _)
-              (delete-char (- 0 (length str)))
-              (insert (concat (unless roam-p "roam:")
-                              str))
-              (forward-char 2))))))
+              (let ((node-id (org-roam-node-id
+                              (org-roam-node-from-title-or-alias
+                               (substring-no-properties str)))))
+                (cond (old-id (delete-char
+                               (- 0 (length str) (length old-id) 2))
+                              (insert node-id "][" str))
+                      (t (delete-char (- 0 (length str)))
+                         (insert "id:" node-id "][" str)))
+                (forward-char 2)))))))
 
 (defun org-roam-complete-everywhere ()
   "Complete symbol at point as a link completion to an Org-roam node.
@@ -788,7 +796,7 @@ This is a `completion-at-point' function, and is active when
 `org-roam-completion-everywhere' is non-nil.
 
 Unlike `org-roam-complete-link-at-point' this will complete even
-outside of the bracket syntax for links (i.e. \"[[roam:|]]\"),
+outside of the bracket syntax for links (i.e. \"[[id:|]]\"),
 hence \"everywhere\"."
   (when (and org-roam-completion-everywhere
              (thing-at-point 'word)
@@ -799,7 +807,10 @@ hence \"everywhere\"."
             :exit-function
             (lambda (str _status)
               (delete-char (- (length str)))
-              (insert "[[roam:" str "]]"))
+              (insert "[[id:"
+                      (org-roam-node-id (org-roam-node-from-title-or-alias
+                                         (substring-no-properties str)))
+                      "][" str "]]"))
             ;; Proceed with the next completion function if the returned titles
             ;; do not match. This allows the default Org capfs or custom capfs
             ;; of lower priority to run.